chore: Make a more universal lyrics time string parser
- Should work with `mm:ss.xxx` & `ss.xxx` formats.
  - The `ss.xxx` format will appear with TTML lyrics.
